% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_ais.R
\name{downsample_ais}
\alias{downsample_ais}
\title{Returns null, just processes data from web}
\usage{
downsample_ais(df, every_minutes = 10, status_codes_to_keep = c(0, 7, 8, 9,
  10, 11, 12, 13, 14, 15), SOG_threshold = 1, vessel_attr = c("VesselType",
  "Length"), voyage_attr = c("Destination"), raw = FALSE)
}
\arguments{
\item{df}{The data frame consisting of the combinations of zones, years, months to process}

\item{every_minutes}{How often algorithm finds the closest points to (e.g. 10 minute, 30 minute, etc)}

\item{status_codes_to_keep}{Status codes to retain for downsampling. List here: https://help.marinetraffic.com/hc/en-us/articles/203990998-What-is-the-significance-of-the-AIS-Navigational-Status-Values-}

\item{SOG_threshold}{threshold to to delete values less than this (corresponding to little or no movement)}

\item{vessel_attr}{The attributes from the vessel table to join in}

\item{voyage_attr}{The attributes from the voyage table to join in}

\item{raw}{If true, no filtering based on time applied. Defaults to FALSE.}
}
\description{
Returns null, just processes data from web
}
