% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxDepth.R
\name{maxDepth}
\alias{maxDepth}
\title{Find the maximum depth for every whale & vessel pair}
\usage{
maxDepth(df, map, depth, segment_length = 0.5)
}
\arguments{
\item{df}{The data frame that join whale and vessel together. Necessary variables: whale's loc (UTM), vessels'loc (UTM)}

\item{map}{The map that around whale and vessel}

\item{depth}{The depth data include location (x,y) and depth (in meters)}

\item{segment_length}{in km}
}
\value{
maximum depth for each whale & vessel pair
}
\description{
Find the maximum depth for every whale & vessel pair
}
