% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ais_data.R
\name{process_ais_data}
\alias{process_ais_data}
\title{Processes raw AIS data, creates cleaned data frame of hourly observations with speed calculations}
\usage{
process_ais_data(aisdata, min_from_hr = 7.5, time_zone = "GMT")
}
\arguments{
\item{aisdata}{The data frame consisting of the combinations of zones, years, months to process}

\item{min_from_hr}{Window around the top of the hour used. Every vessel might not have a position recorded at exactly 12:00 for example, so the closest position +/- minutes_from_hour is used. Defaults to 7.5 (15 min window)}

\item{time_zone}{The time zone used in calculations (e.g. "GMT", "UTC")}
}
\value{
processed data frame
}
\description{
Processes raw AIS data, creates cleaned data frame of hourly observations with speed calculations
}
\examples{
\dontrun{
df = data.frame("month"=1:4, "year" = 2009, "zone"=10)
downsample_ais(df, raw = TRUE) # gets raw data from marine cadastre

d = readRDS("filtered/Zone10_01_2009.rds")
d_processed = process_ais_data(d, time_zone = "UTC")
}
}
