% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRL.R
\name{getRL}
\alias{getRL}
\title{Returns a dataframe that includes date, Received Level(dB) and location of the whale}
\usage{
getRL(whale, vessel)
}
\arguments{
\item{whale}{The data frame include date(PST), position(UTM) of the whale}

\item{vessel}{The data frame include date(PST), position(UTM) of the vessel}
}
\value{
dataframe
}
\description{
Returns a dataframe that includes date, Received Level(dB) and location of the whale
}
